# Avoid unnecessary copy-on-write Vector/Array