--- a/core/templates/vector.h
+++ b/core/templates/vector.h
@@ -315,8 +315,8 @@ public:
 
 template <typename T>
 void Vector<T>::reverse() {
+	T *p = ptrw();
 	for (Size i = 0; i < size() / 2; i++) {
-		T *p = ptrw();
 		SWAP(p[i], p[size() - i - 1]);
 	}
 }
@@ -329,8 +329,9 @@ void Vector<T>::append_array(Vector<T> p_other) {
 	}
 	const Size bs = size();
 	resize(bs + ds);
+	T *p = ptrw();
 	for (Size i = 0; i < ds; ++i) {
-		ptrw()[bs + i] = p_other[i];
+		p[bs + i] = p_other[i];
 	}
 }