<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->

There are some `ptrw` used in `Array` and `Vector` where read-only access is needed. Using `ptr` to avoid copy-on-write. There were also some loops that called `ptrw` so I made them similar to other loops that create `ptrw` once outside the loop.