@Calinou 2024-06-17
Are there any benchmarks to showcase a performance difference?