@Mickeon 2024-06-17
Making `bsearch()` constant causes the check to fail because no compatibility method was provided as well @AThousandShips