@mashumafi 2024-06-17
> Making `bsearch()` constant causes the check to fail because no compatibility method was provided as well @AThousandShips

Should I avoid changing the signature to be `const`?



> Are there any benchmarks to showcase a performance difference?

I created an ad-hoc benchmark using `reverse` to get an idea.

Results:

| Operation     | Element Count | master timing | PR time       |
| ------------- | ------------- | ------------- | ------------- |
| reverse       | 100           | 0.012         | 0.012         |
| reverse       | 10,000        | 0.046         | 0.038         |
| reverse       | 1,000,000     | 3.716         | 3.096         |

```gdscript
extends Benchmark

static func create_array(size: int) -> Array[int]:
	var result : Array[int] = []
	result.resize(size)
	for i in size:
		result[i] = i
	return result

var array_100 := create_array(100)
var array_dup_100 := array_100

var array_10_000 := create_array(10_000)
var array_dup_10_000 := array_10_000

var array_1_000_000 := create_array(1_000_000)
var array_dup_1_000_000 := array_1_000_000

func benchmark_reverse_100() -> void:
	array_100.reverse()

func benchmark_reverse_10_000() -> void:
	array_10_000.reverse()

func benchmark_reverse_1_000_000() -> void:
	array_1_000_000.reverse()

```