@mashumafi 2024-06-18
I added these functions to https://github.com/godotengine/godot-benchmarks/pull/75