@mashumafi 2024-06-18
> This is a duplicate of:
> 
> * [Make `Vector.bsearch` and `bsearch_custom` const #90341](https://github.com/godotengine/godot/pull/90341)
> 
> See [#79140 (comment)](https://github.com/godotengine/godot/pull/79140#discussion_r1417289615) for the issues with this (we can't handle compatibility breakages with these types right now)

I reverted the `const` portion of the change. I fell back to the original intent of the PR which is to move away from `ptrw` where it makes sense. Hoping this means the `breaks compat` tag can be removed and considered to be merged again.