@AThousandShips 2024-06-18
You broke your rebase and you now have a merge commit in your branch, which is not allowed, please reset your branch with `git reset --hard master` and then reapply your changes, the current branch won't work