@monxa 2024-06-20
Hi! I am the OP of the other PR concerning the same issue. I do agree with the assessment here.

Given we have a variant object which contains a template-vector-type, e.g., PackedXXXArray, if we want to `bsearch` that, we'd have to use it plainly without reference, cast that to the vector type, which does trigger COW.

So we need the ability to use it by reference. Having `bsearch` const is important. Changing `ptrw` to `ptr` does not address the issue, unfortunately.

We need to provide backwards compatibility because the method hashes depend on the constness of the method signature (
https://github.com/godotengine/godot/blob/b75f0485ba15951b87f1d9a2d8dd0fcd55e178e4/core/variant/variant_call.cpp#L1402-L1420
).

Now, I will find time in the next weeks to go back and look at compatibility for GDExtension for variant_calls/global methods. If I remember correctly, there is a way to catch the check in gdextension.cpp without impacting the performance of the current state. After research, I know it can be done, and I even know somewhat how. But it is a very complicated undertaking (at least for me).

Again, if someone wants a real challenge, please consider providing a compatibility layer for variant_calls. I would also be open to collaboratively working on that.

If such a compatibility layer is not provided, important changes like this will not be merged until we get the next major release, which will take YEARS.

Here are my raw notes from last time I worked on this:
[monxa_notes_internal_compat_binds.md](https://github.com/user-attachments/files/15910676/monxa_notes_internal_compat_binds.md)