@RandomShaper 2024-06-21
This one still deals with `reverse()`, so it's at least not a strict subset of #90341 and it's so simple that doesn't look risky at all.